move metrics => cli
