+++ cli/metrics/client.go
