+++ cli/metrics/commands.go
