+++ cli/metrics/generatecommands/main.go
