+++ server/metrics.go
@@ -21,7 +21,7 @@ import (
 
 	"google.golang.org/grpc"
 
-	"github.com/docker/compose-cli/metrics"
+	"github.com/docker/compose-cli/cli/metrics"
 	"github.com/docker/compose-cli/server/proxy"
 )
