+++ server/server.go
@@ -25,7 +25,7 @@ import (
 	"google.golang.org/grpc/health"
 	"google.golang.org/grpc/health/grpc_health_v1"
 
-	"github.com/docker/compose-cli/metrics"
+	"github.com/docker/compose-cli/cli/metrics"
 )
 
 // New returns a new GRPC server.
